refactor: move useModalAnimation hook to own file
it is a react best practice to keep hooks in their own file in a /hooks folder
